+++ tests/hazmat/primitives/test_hash_vectors.py
@@ -101,6 +101,40 @@ class TestSHA512(object):
 
 
 @pytest.mark.supported(
+    only_if=lambda backend: backend.hash_supported(hashes.SHA512_224()),
+    skip_message="Does not support SHA512/224",
+)
+@pytest.mark.requires_backend_interface(interface=HashBackend)
+class TestSHA512224(object):
+    test_SHA512_224 = generate_hash_test(
+        load_hash_vectors,
+        os.path.join("hashes", "SHA2"),
+        [
+            "SHA512_224LongMsg.rsp",
+            "SHA512_224ShortMsg.rsp",
+        ],
+        hashes.SHA512_224(),
+    )
+
+
+@pytest.mark.supported(
+    only_if=lambda backend: backend.hash_supported(hashes.SHA512_256()),
+    skip_message="Does not support SHA512/256",
+)
+@pytest.mark.requires_backend_interface(interface=HashBackend)
+class TestSHA512256(object):
+    test_SHA512_256 = generate_hash_test(
+        load_hash_vectors,
+        os.path.join("hashes", "SHA2"),
+        [
+            "SHA512_256LongMsg.rsp",
+            "SHA512_256ShortMsg.rsp",
+        ],
+        hashes.SHA512_256(),
+    )
+
+
+@pytest.mark.supported(
     only_if=lambda backend: backend.hash_supported(hashes.MD5()),
     skip_message="Does not support MD5",
 )